requirements.txt
aiomysql==0.0.7
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
flake8==3.6.0
idna==2.8
lxml==4.2.5
mccabe==0.6.1
msgpack==0.6.0
numpy==1.15.4
pandas==0.23.4
pycodestyle==2.4.0
pyflakes==2.0.0
PyMySQL==0.6.7
python-dateutil==2.7.5
pytz==2018.7
pyzmq==17.1.2
requests==2.21.0
selenium==3.141.0
simplejson==3.16.0
six==1.11.0
SQLAlchemy==1.2.14
tushare==1.2.18
urllib3==1.24.1
wincertstore==0.2
yapf==0.25.0
